make layerToSelection work with multiple of layers
